Refactor code structure for improved readability and maintainability

+++ backend/app/models/saisie_temps.py
@@ -1,6 +1,6 @@
 # app/models/saisie_temps.py
 
-from sqlalchemy import Column, Integer, Date, DateTime, ForeignKey, String, Float, CheckConstraint
+from sqlalchemy import Column, Integer, Date, DateTime, ForeignKey, String, CheckConstraint
 from sqlalchemy.sql import func
 from sqlalchemy.orm import relationship
 
@@ -12,7 +12,7 @@ class SaisieTemps(Base):
     id = Column(Integer, primary_key=True, index=True)
     tache_id = Column(Integer, ForeignKey("tache.id", ondelete="CASCADE"), nullable=False, index=True)
     utilisateur_id = Column(Integer, ForeignKey("utilisateur.id", ondelete="CASCADE"), nullable=False, index=True)
-    duree_min = Column(Float, nullable=False)
+    duree_min = Column(Integer, nullable=False)
     date = Column(Date, nullable=False, server_default=func.current_date())
     note = Column(String(500), nullable=True)
     cree_le = Column(DateTime(timezone=True), server_default=func.now())
